requirements.txt
requests~=2.32.5
Flask~=3.1.2
python-dotenv~=1.2.1